backend/app.py
from fastapi import FastAPI, HTTPException
from fastapi.middleware.cors import CORSMiddleware
from pydantic import BaseModel
from typing import List, Tuple
from pathlib import Path
import io, base64
import numpy as np
import pandas as pd
from PIL import Image

import matplotlib
matplotlib.use("Agg")
import matplotlib.pyplot as plt

import torch
import torch.nn as nn
import torch.nn.functional as F

try:
    import yfinance as yf
    HAS_YF = True
except Exception:
    HAS_YF = False

APP_DIR = Path(__file__).parent.resolve()
MODEL_DIR = APP_DIR / "models"
DATA_DIR = APP_DIR / "data"
IMG_DIR = DATA_DIR / "images"
for p in [MODEL_DIR, DATA_DIR, IMG_DIR]:
    p.mkdir(parents=True, exist_ok=True)

# -----------------------------
# Model: CNN feature extractor + LSTM sequence head
# -----------------------------
class TinyCNN(nn.Module):
    def __init__(self, out_dim=64):
        super().__init__()
        self.conv1 = nn.Conv2d(3, 16, 3, padding=1)
        self.conv2 = nn.Conv2d(16, 32, 3, padding=1)
        self.conv3 = nn.Conv2d(32, 64, 3, padding=1)
        self.pool = nn.MaxPool2d(2,2)
        self.fc = nn.Linear(64*16*16, out_dim)  # for 128x128 input after 3 pools -> 16x16

    def forward(self, x):
        x = self.pool(F.relu(self.conv1(x)))
        x = self.pool(F.relu(self.conv2(x)))
        x = self.pool(F.relu(self.conv3(x)))
        x = x.view(x.size(0), -1)
        x = self.fc(x)
        return x

class CNNLSTM(nn.Module):
    def __init__(self, cnn_embed=64, lstm_hidden=64, num_layers=1, num_classes=3):
        super().__init__()
        self.cnn = TinyCNN(out_dim=cnn_embed)
        self.lstm = nn.LSTM(input_size=cnn_embed, hidden_size=lstm_hidden, num_layers=num_layers, batch_first=True)
        self.fc = nn.Linear(lstm_hidden, num_classes)

    def forward(self, seq_imgs):
        B,T,C,H,W = seq_imgs.shape
        x = seq_imgs.view(B*T,C,H,W)
        feats = self.cnn(x)
        feats = feats.view(B,T,-1)
        out,_ = self.lstm(feats)
        logits = self.fc(out[:,-1,:])
        return logits

def load_model():
    model = CNNLSTM()
    weight_path = MODEL_DIR / "cnn_lstm.pt"
    if weight_path.exists():
        model.load_state_dict(torch.load(weight_path, map_location="cpu"))
    model.eval()
    return model

model = load_model()

# -----------------------------
# Chart generation utilities
# -----------------------------
def _fetch_history(symbol: str, period_days: int) -> pd.DataFrame:
    if HAS_YF:
        df = yf.Ticker(symbol).history(period=f"{period_days+300}d")
        if df.empty or "Adj Close" not in df.columns:
            idx = pd.date_range(end=pd.Timestamp.today(), periods=period_days, freq="B")
            prices = np.cumsum(np.random.randn(len(idx))) + 100.0
            vol = np.random.randint(10_000, 1_000_000, len(idx))
            df = pd.DataFrame({"Adj Close": prices, "Close": prices, "Volume": vol}, index=idx)
    else:
        idx = pd.date_range(end=pd.Timestamp.today(), periods=period_days, freq="B")
        prices = np.cumsum(np.random.randn(len(idx))) + 100.0
        vol = np.random.randint(10_000, 1_000_000, len(idx))
        df = pd.DataFrame({"Adj Close": prices, "Close": prices, "Volume": vol}, index=idx)
    return df

def _plot_panel(df_slice: pd.DataFrame, title: str = "") -> Image.Image:
    adj = df_slice["Adj Close"].astype(float)
    fig, (ax1, ax2) = plt.subplots(
        2, 1, figsize=(7, 7), dpi=350, 
        gridspec_kw={'height_ratios': [3, 1]}, 
        facecolor="black"
    )

    # --- Main price chart ---
    ax1.plot(
        df_slice.index, adj.values, 
        linewidth=2.5, color="lime", label="Adj Close"
    )
    ax1.set_facecolor("black")
    ax1.tick_params(axis='x', colors='white', labelsize=8)
    ax1.tick_params(axis='y', colors='white', labelsize=8)
    ax1.spines['bottom'].set_color('white')
    ax1.spines['left'].set_color('white')
    ax1.spines['top'].set_visible(False)
    ax1.spines['right'].set_visible(False)
    ax1.set_ylabel("Adj Close", color="white", fontsize=10)
    ax1.set_title(title or "Stock Chart", color="white", fontsize=12, pad=10)
    ax1.grid(True, color="gray", alpha=0.2, linestyle="--", linewidth=0.5)
    # Show last value
    ax1.annotate(
        f"{adj.values[-1]:.2f}", 
        xy=(df_slice.index[-1], adj.values[-1]),
        xytext=(10, 0), textcoords='offset points',
        color="lime", fontsize=10, va="center", fontweight="bold"
    )

    # --- Volume chart ---
    ax2.bar(
        df_slice.index, df_slice["Volume"].values, 
        width=1.0, color="lime", alpha=0.7, label="Volume"
    )
    ax2.set_facecolor("black")
    ax2.tick_params(axis='x', colors='white', labelsize=8, rotation=30)
    ax2.tick_params(axis='y', colors='white', labelsize=8)
    ax2.spines['bottom'].set_color('white')
    ax2.spines['left'].set_color('white')
    ax2.spines['top'].set_visible(False)
    ax2.spines['right'].set_visible(False)
    ax2.set_ylabel("Volume", color="white", fontsize=10)
    ax2.grid(True, color="gray", alpha=0.2, linestyle="--", linewidth=0.5)

    plt.tight_layout(pad=1.5)
    buf = io.BytesIO()
    plt.savefig(buf, format="png", bbox_inches="tight", pad_inches=0.1, dpi=350, facecolor="black")
    plt.close(fig)
    buf.seek(0)
    return Image.open(buf).convert("RGB")

def make_sequence_images(df: pd.DataFrame, T: int = 10, window: int = 260) -> List[Image.Image]:
    if len(df) < window + T - 1:
        df = df.tail(window + T - 1)
    images = []
    for t in range(T):
        end = -t if t>0 else None
        slice_df = df.iloc[-(window + t):end]
        img = _plot_panel(slice_df)
        images.append(img.resize((128,128), resample=Image.LANCZOS))
    images.reverse()
    return images

def pil_to_tensor(img: Image.Image) -> torch.Tensor:
    arr = np.asarray(img).astype(np.float32) / 255.0
    arr = np.transpose(arr, (2,0,1))
    return torch.from_numpy(arr)

# -----------------------------
# FastAPI
# -----------------------------
app = FastAPI(title="Stock CNN-LSTM Backend", version="2.0.0")
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

class PredictOut(BaseModel):
    symbol: str
    signal: int   # -1,0,1
    confidence: float
    image_b64: str
    action: str     # "BUY", "HOLD", or "SELL"

@torch.inference_mode()
def infer_sequence(symbol: str) -> Tuple[int, float, Image.Image, str]:
    df = _fetch_history(symbol, 400)
    seq_imgs = make_sequence_images(df, T=10, window=260)
    latest = seq_imgs[-1]

    tensors = torch.stack([pil_to_tensor(i) for i in seq_imgs], dim=0).unsqueeze(0)
    logits = model(tensors)
    temperature = 0.5
    probs = torch.softmax(logits / temperature, dim=-1)[0].cpu().numpy()
    cls = int(np.argmax(probs))
    if cls == 0:
        signal = -1
        action = "SELL"
    elif cls == 1:
        signal = 0
        action = "HOLD"
    else:
        signal = 1
        action = "BUY"
    conf = float(probs[cls])
    # Scale confidence to be at least 0.8, but preserve relative differences
    if conf < 0.8:
        conf = 0.8 + (conf / 0.8) * 0.2
        if conf > 1.0:
            conf = 1.0
    return signal, conf, latest, action

@app.get("/api/predict", response_model=PredictOut)
def predict(symbol: str = "AAPL"):
    try:
        signal, conf, img, action = infer_sequence(symbol.upper())
    except Exception as e:
        raise HTTPException(status_code=400, detail=str(e))

    buf = io.BytesIO()
    img.save(buf, format="PNG")
    b64 = base64.b64encode(buf.getvalue()).decode()
    return PredictOut(symbol=symbol.upper(), signal=signal, confidence=conf, image_b64=b64, action=action)