backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2
pillow==10.4.0
matplotlib==3.9.2
numpy==2.0.1
pandas==2.2.2
yfinance==0.2.43

torch==2.3.1
